Automatically delete/create the index
This turns the cell that creates the new index (deleting it if it
was already present) into a code cell.

Having it as a Markdown cell was confusing, and in practice
everyone whose uses this notebook had to first turn it into a code
cell, or write a code cell with similar code.

Automatically deleting and creating the index is also more in line
with how the other notebooks showcasing Pinecone work.