Use the path returned by wget.download
Usually the download path will be the hard-coded value being used
before, but not always. If a file exists there, wget will choose
another filename (with a parenthesized number in the basename). If
this is a redundant redownloading, then it doesn't matter, but if
an old, corrupted, or otherwise different version was downloaded
there before, it would be used again instead of the new file.