Use relative links between notebooks/documents
This repository already usually does that, but there were some
cases where it was not done and did not seem intentional. This
fixes those.

In most cases, relative links work best, because they work both
locally and in editors, and when they are links to notebooks, they
open up the notebooks properly instead of opening an external
website.

There are some exceptions where it is best to keep them as absolute
URLs, such as code in modules that not all editors will linkify,
links from notebooks to directories, and links that would become
wrong if ever resolved elsewhere, such as those related to the
issue template and discussion pages. This leaves those alone.