Replace broken Langchain documentation links
The Langchain documentation has changed and been reorganized over
time, and a number of the links in "How to build a tool-using agent
with Langchain" have broken.

This replaces the links to pages that are no longer part of the
Langchain documentation, substituting links to pages in the current
documentation that cover the relevant material. In most cases, a
strong, close replacement was available.

Similar replacements may need to be done in some other notebooks,
but this specific change only modifies that one notebook.

The notebook also included a number of "weakly broken" links, going
to a 404 page but being immediately redirected to a relevant page.
This specific change doesn't modify those; I'll do that separately.

Although the relevance of all the replacements can be checked
efficiently, it is not always obvious how they were identified. The
following notes clarify that in case it needs to be illuminated
for troubleshooting purposes, but can otherwise be ignored. (Note
that this change doesn't replace any links with archive.org copies,
but I give those where available in my notes below.)

https://python.langchain.com/docs/modules/agents/agents/examples/mrkl_chat.html
https://web.archive.org/web/20230609050957/https://python.langchain.com/en/latest/modules/agents/agents/examples/mrkl_chat.html
Clear best replacement (covers both non-chat and chat):
https://python.langchain.com/docs/modules/agents/how_to/mrkl

https://python.langchain.com/en/latest/modules/agents/agents/custom_llm_agent.html
Not archived.
Two very good candidates for the best replacement:
https://python.langchain.com/docs/modules/agents/how_to/custom_llm_agent
https://python.langchain.com/docs/modules/agents/how_to/custom_llm_chat_agent (chosen)

https://python.langchain.com/en/latest/modules/prompts/prompt_templates.html
https://web.archive.org/web/20230610132710/https://python.langchain.com/en/latest/modules/prompts/prompt_templates.html
Clear best replacement:
https://python.langchain.com/docs/modules/model_io/prompts/prompt_templates/

https://python.langchain.com/docs/modules/agents/agents/custom_llm_agent.html#adding-memory
Not archived (see above).
No perfect candidate for the best replacement, but two reasonable ones:
https://python.langchain.com/docs/modules/memory/how_to/agent_with_memory
https://python.langchain.com/docs/modules/memory/how_to/buffer_window (chosen)

https://api.python.langchain.com/en/latest/modules/vectorstore.html
Not archived.
Clear best replacement (because we are using langchain.vectorstores):
https://api.python.langchain.com/en/latest/api_reference.html#module-langchain.vectorstores

https://python.langchain.com/docs/modules/chains/examples/sqlite.html
Not archived.
Clear best replacement:
https://python.langchain.com/docs/modules/chains/popular/sqlite

https://python.langchain.com/docs/modules/chains/index_examples/vector_db_qa.html
Not archived.
Pretty good replacement:
https://python.langchain.com/docs/modules/chains/additional/openai_functions_retrieval_qa