Update working Langchain documentation links
This improves links in the "How to build a tool-using agent with
Langchain" notebook that are currently redirects (some of which
briefly go to a 404 page but then automatically redirect to
a relevant document).

This is not as important as fixing the broken links, but I think
it makes them less likely to break in the future (at least for the
ones that were redirecting off a 404 page), and in the case of most
or all of them, makes loading the ultimate target page faster.