Link "Learn more" for "user" argument to the guide
In one of the DALL-E notebooks, the old beta.openai.com link for
information about the "user" argument (used for monitoring) had
once pointed to a page that documented it. But it came to be
redirected to the newer "Usage policies" page, which makes no
reference to that argument. This changes it to point to the
"End-user IDs" section of the "Safety best practices" guide,
instead.