Update "Zero-Shot Reasoners" PDF link
This removes the "v1" at the end of the direct PDF link to the
"Large Language Models are Zero-Shot Reasoners" paper on arXiv.
There are currently four versions of the paper, and linking
specifically to v1 appears unintentional. This now links to v4, and
will automatically update if further updates to the paper are made
(such as updating authors' institutional affiliations).