Fix citations where "et al." stood for one author
The "Faithful Reasoning" paper is a sequel to the
"Selection-Inference" paper, but one of the "Selection-Inference"
authors is not an author of the "Faithful Reasoning" paper. As a
result, the "Faithful Reasoning" paper has only two authors, and
"et al." in the citations to it was standing just for the paper's
second author.

This replaces that with the second author's name, while leaving
citations to the three-author "Selection-Inference" paper, as well
as the description of the relationship between the two papers,
unchanged.