Fix a few more Langchain documentation links
The vector stores link redirected off the 404 page to a page about
vector store, but that new page did not have the list of supported
vector stores that it was being linked to for, so I've changed it
to the current page that has that.

The databases page link was fully broken. I've pointed it to the
corresponding new page. It seems one goal of this link is to get a
list of different supported databases, which the curren page does
not have, but browsing on archive.org suggests the old page didn't
have that either.