Fix davinci-instruct-beta-v3 link
The olympics-2-create-qa notebook has several occurrences of
davinci-insruct-beta-v3 that are links to the old documentation on
beta.openai.com. This is the old name for text-davinci-002, but the
old link now redirects to https://platform.openai.com/docs/models.
Neither that nor https://platform.openai.com/docs/deprecations
mention davinci-instruct-beta-v3. So I have changed these links to
https://help.openai.com/en/articles/5832130-what-s-changed-with-engine-names-and-best-practices#h_e19b74061f.

One alternative would be to change the name to text-davinci-001 and
link to the deprecated model page. But elsewhere in the notebook it
often talks about "davinci instruct", and the code in the notebook
itself refers to davinci-instruct-beta-v3. So if this were done,
then further changes would have to be made to preserve clarity.
Changing the links to the remaining page on openai.com that clearly
states what that model name means is the simplest way to fix the
problem.