Use relative link in Chat finetuning data prep
This changes the Markdown links to the data file to a relative
link, preserving its treatment by browsers while also making
editors like VS Code open it in them (rather than an external
browser) when clicked.

This doesn't change code or even comments in that notebook, just
the link in a Markdown cell.