More consistent Kusto doc link style
This changes one link to the Microsoft Learn documentation to
remove "?tabs=query-defined", which is not specified elsewhere, is
automatically added when visiting the site, and that is probably
not specifically preferred (so if the default "tabs" value changes,
there is no reason to override it).

While I'm at it, I've also removed a spurious blank leading cell in
the notebooks, and tweaked style of the associated readme file to
make it clearer in ways it differs from the prevailing style of
other .md files in this repository.