Fix and update links in Redis readme
This fixes two links, one of which was malformed (it didn't have
a protocol or domain name), and the other of which used to work but
now 404s.

This also update a number of other links in the same file to their
redirect targets. Most of the old links are unlikely to break, but
in this case updating them does make the pages load a bit faster.